requirements.txt
numpy==1.19.2
tensorflow-gpu==2.4.0
python_speech_features==0.6
pydub==0.25.1
scipy==1.7.1
jiwer==2.2.0
click==8.0.3
tqdm==4.62.2
matplotlib==3.4.2
pandas==1.3.1
pypinyin==0.43.0